requirements.txt

streamlit>=1.15.0
openai>=1.0.0
pandas>=2.0
networkx>=3.0
matplotlib>=3.4
python-dotenv>=1.0.0
